move AddressData, Routedata to settings/datatypes.py
