+++ sdbus_async/networkmanager/settings/datatypes.py
@@ -1,0 +1,32 @@
+from dataclasses import dataclass, field
+from typing import Optional
+
+from .base import NetworkManagerSettingsMixin
+
+
+@dataclass
+class AddressData(NetworkManagerSettingsMixin):
+    address: str = field(
+        metadata={'dbus_name': 'address', 'dbus_type': 's'},
+    )
+    prefix: int = field(
+        metadata={'dbus_name': 'prefix', 'dbus_type': 'u'},
+    )
+
+
+@dataclass
+class RouteData(NetworkManagerSettingsMixin):
+    dest: str = field(
+        metadata={'dbus_name': 'dest', 'dbus_type': 's'},
+    )
+    prefix: int = field(
+        metadata={'dbus_name': 'prefix', 'dbus_type': 'u'},
+    )
+    next_hop: Optional[str] = field(
+        metadata={'dbus_name': 'next-hop', 'dbus_type': 's'},
+        default=None,
+    )
+    metric: Optional[int] = field(
+        metadata={'dbus_name': 'metric', 'dbus_type': 'u'},
+        default=None,
+    )
